+++ src/main/java/domain/BonusDraw.java
@@ -1,4 +1,4 @@
-package service;
+package domain;
 
 import lotto.Validation;
