+++ src/main/java/domain/LottoDraw.java
@@ -1,4 +1,4 @@
-package service;
+package domain;
 
 import lotto.Validation;
